One sentence per line and miscellaneous spelling and formatting changes
Some paragraphs in the source text were on a single line.  This is
sometimes difficult to read but it also makes it difficult to see
the actual changes between versions when displayed in some diff
programmes.

I have split multiple sentence lines into several single sentence lines
accordingly.  A further step might be to split the clauses in the
sentences onto single lines.

Furthermore, I made a number of changes as I encountered problems in
processing the source:
	- Changed the hyperref package call in ElisaPaper.cls to
	eliminate the toc entries being enclosed in red boxes
	- Changed the form of the telco times tables
	- Added hyphens to all instances of "safety critical" and
	"safety relevant"
	- Corrected miscellansous spelling mistakes as detected by
	the spelling checker (US American spelling... ugh)

Signed-off-by: John MacGregor <open.john.macgregor@gmail.com>